+++ tests/index.test.js
@@ -1,33 +1,8 @@
-const fs = require('fs');
-const unified = require('unified');
-const markdown = require('remark-parse');
-const html = require('remark-html');
-const HBS = require('../index');
-const { parse } = require('ember-template-recast');
-const chalk = require('chalk');
-const emphasize = require('emphasize');
 const { stripIndent } = require('common-tags');
-
-const stack = unified().use(markdown).use(HBS).use(html);
-
-const process = code => stack.processSync(code).toString();
-
-let files = [`/fixtures/example.md`, '/fixtures/nested-namespaced.md'];
-
-for (let filePath of files) {
-  test(`${filePath}: it should keep all the HBS and HTML`, () => {
-    const file = stack.processSync(fs.readFileSync(__dirname + filePath));
-
-    const contents = file.toString();
-
-    expect(contents).toMatchSnapshot();
-
-    verifyContents(contents);
-  });
-}
+const { verifyContents, transform } = require('./utils');
 
 test('can render text in a block component', () => {
-  const contents = process(stripIndent`
+  const contents = transform(stripIndent`
     <Nested::Component>
       some text
     </Nested::Component>
@@ -41,22 +16,3 @@ test('can render text in a block component', () => {
 
   verifyContents(contents);
 });
-
-// Validate that the contents are parseable
-// this will throw an exception if they are not
-function verifyContents(code) {
-  try {
-    parse(code);
-  } catch (e) {
-    console.log(
-      chalk.red(e.message) +
-        '\n\n' +
-        emphasize.highlight('handlebars', code).value +
-        chalk.red(
-          '\n' +
-            `^ ----- contents after [ markdown -> HBS -> html ] ----- ^`
-        )
-    );
-    throw e;
-  }
-}
